# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MissionLedger API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/profit-loss:
    get:
      operationId: getProfitLossReport
      tags: [reports]
      summary: Get profit and loss report
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: fundId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: P&L report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitLossReport"
  /reports/balance-sheet:
    get:
      operationId: getBalanceSheetReport
      tags: [reports]
      summary: Get balance sheet report
      parameters:
        - name: asOfDate
          in: query
          schema:
            type: string
        - name: fundId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Balance sheet report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceSheetReport"
  /reports/cash-flow:
    get:
      operationId: getCashFlowReport
      tags: [reports]
      summary: Get cash flow report
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Cash flow report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashFlowReport"
  /reports/budget-vs-actual:
    get:
      operationId: getBudgetVsActualReport
      tags: [reports]
      summary: Get budget vs actual report
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: budgetId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Budget vs actual report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetVsActualReport"
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
...
  /reconciliation:
    get:
      operationId: getReconciliations
      tags: [reconciliation]
      summary: Get all reconciliations
      responses:
        "200":
          description: List of reconciliations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reconciliation"
    post:
      operationId: createReconciliation
      tags: [reconciliation]
      summary: Create a reconciliation session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReconciliationRequest"
      responses:
        "201":
          description: Reconciliation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reconciliation"
  /reconciliation/{id}/items:
    get:
      operationId: getReconciliationItems
      tags: [reconciliation]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reconciliation items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReconciliationItem"
    put:
      operationId: updateReconciliationItems
      tags: [reconciliation]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                itemIds:
                  type: array
                  items:
                    type: string
                cleared:
                  type: boolean
              required: [itemIds, cleared]
      responses:
        "200":
          description: Updated
  /reconciliation/{id}/finish:
    post:
      operationId: finishReconciliation
      tags: [reconciliation]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reconciliation finished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reconciliation"
components:
  schemas:
...
    Fund:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        balance:
          type: number
        totalDonations:
          type: number
        totalExpenses:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, isActive, createdAt, updatedAt]
...
    Account:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        type:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        parentId:
          type: string
          nullable: true
        balance:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, code, name, type, isActive, createdAt, updatedAt]
...
    BankTransaction:
      type: object
      properties:
        id:
          type: string
        bankAccountId:
          type: string
        date:
          type: string
        description:
          type: string
        merchantName:
          type: string
          nullable: true
        amount:
          type: number
        type:
          type: string
        status:
          type: string
        fundId:
          type: string
          nullable: true
        accountId:
          type: string
          nullable: true
        account:
          $ref: "#/components/schemas/Account"
          nullable: true
        fund:
          $ref: "#/components/schemas/Fund"
          nullable: true
        journalEntryId:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, bankAccountId, date, description, amount, type, status, createdAt, updatedAt]
...
    ProfitLossReport:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        revenue:
          type: array
          items:
            $ref: "#/components/schemas/ReportLineItem"
        expenses:
          type: array
          items:
            $ref: "#/components/schemas/ReportLineItem"
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        netIncome:
          type: number
      required: [startDate, endDate, revenue, expenses, totalRevenue, totalExpenses, netIncome]
    BalanceSheetReport:
      type: object
      properties:
        asOfDate:
          type: string
        assets:
          type: array
          items:
            $ref: "#/components/schemas/ReportLineItem"
        liabilities:
          type: array
          items:
            $ref: "#/components/schemas/ReportLineItem"
        equity:
          type: array
          items:
            $ref: "#/components/schemas/ReportLineItem"
        totalAssets:
          type: number
        totalLiabilities:
          type: number
        totalEquity:
          type: number
      required: [asOfDate, assets, liabilities, equity, totalAssets, totalLiabilities, totalEquity]
    CashFlowReport:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        operating:
          type: array
          items:
            $ref: "#/components/schemas/ReportLineItem"
        totalCashFlow:
          type: number
      required: [startDate, endDate, operating, totalCashFlow]
    BudgetVsActualReport:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              accountName:
                type: string
              budgeted:
                type: number
              actual:
                type: number
              variance:
                type: number
              variancePercent:
                type: number
            required: [accountName, budgeted, actual, variance, variancePercent]
      required: [startDate, endDate, items]
    ReportLineItem:
      type: object
      properties:
        accountId:
          type: string
        accountCode:
          type: string
        accountName:
          type: string
        amount:
          type: number
        children:
          type: array
          items:
            $ref: "#/components/schemas/ReportLineItem"
      required: [accountId, accountCode, accountName, amount]
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, email, role, isActive, createdAt, updatedAt]
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
      required: [email, password, role]
...
    Reconciliation:
      type: object
      properties:
        id:
          type: string
        bankAccountId:
          type: string
        statementDate:
          type: string
        statementBalance:
          type: number
        status:
          type: string
        clearedBalance:
          type: number
          nullable: true
        difference:
          type: number
          nullable: true
        reconciledAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, bankAccountId, statementDate, statementBalance, status, createdAt, updatedAt]
    CreateReconciliationRequest:
      type: object
      properties:
        bankAccountId:
          type: string
        statementDate:
          type: string
        statementBalance:
          type: number
        openingBalance:
          type: number
      required: [bankAccountId, statementDate, statementBalance]
    ReconciliationItem:
      type: object
      properties:
        id:
          type: string
        reconciliationId:
          type: string
        bankTransactionId:
          type: string
        bankTransaction:
          $ref: "#/components/schemas/BankTransaction"
          nullable: true
        cleared:
          type: boolean
      required: [id, reconciliationId, bankTransactionId, cleared]